Added links and more content
Clarified that lots of this came from the course outline/notes of the pilot course document https://hackmd.io/zEY9rv5QR3O9JFl4jVCYFw?both

I am not 100% clear on the copyright status of that original work, so I'm planning to discuss that with the group. As a result I removed copright notice from this syllabus. I do dedicate my contributions to the public domain, though.